created the sections of the homepage as components for easier code modifications and maintenance
