+++ src/pages/ProjectPage.jsx
@@ -1,6 +1,8 @@
 import React from 'react';
 import { useParams } from 'react-router-dom';
+// Importing component
 import ProjectDetails from '../components/forms/ProjectDetails';
+// Importing projectData
 import projectsData from '../data/projectsData';
 
 function ProjectPage() {
