+++ src/styles/components/typingeffect.css
@@ -1,0 +1,250 @@
+/* ========================================
+========== TYPING-EFFECT-SECTION ==========
+======================================== */
+
+.typing-section {
+    width: 100%;
+    height: auto;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    position: relative;
+    font-size: 1.3rem;
+    padding: 0 0.625rem 3.5rem 0.625rem;
+}
+
+.typing-container {
+    width: 100%;
+    line-height: 1.1;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    gap: 0.5rem;
+}
+
+.typing-group-one {
+    display: flex;
+    align-items: center;
+    justify-content: flex-start;
+    z-index: 1;
+}
+
+.typing-group-two {
+    display: flex;
+    align-items: center;
+    justify-content: flex-end;
+    z-index: 1;
+}
+
+.typing-effect-first {
+    margin-bottom: 0.5rem;
+}
+
+.typing-effect-second {
+    margin-left: 0.5rem;
+}
+
+.typing-effect02 {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    z-index: 1;
+}
+
+.typing-effect::after,
+.typing-effect-first::after,
+.typing-effect-second::after,
+.typing-effect-third::after,
+.typing-effect-fourth::after,
+.typing-effect-fifth::after {
+    content: '';
+    display: inline-block;
+    width: 0.3125rem;
+    height: 1.2rem;
+    background-color: var(--home-bg-color);
+    animation: blink 1s infinite;
+    margin-left: 0.3rem;
+}
+
+@keyframes blink {
+    0%, 100% { opacity: 1; }
+    50% { opacity: 0; }
+}
+
+.book-image,
+.notebook-image {
+    display: flex;
+    flex-direction: row;
+}
+
+.book-image img,
+.notebook-image img {
+    animation: rocking 3s infinite ease-in-out;
+    transform-origin: center top;
+}
+
+.book-image img {
+    width: 4.375rem;
+    height: 5rem;
+    margin-left: 0.5rem;
+    margin-right: 0.5rem;
+}
+
+.notebook-image img {
+    width: 100%;
+    max-width: 5.625rem;
+    height: auto;
+    margin-left: 0.5rem;
+}
+
+@keyframes rocking {
+    0%, 100% { transform: rotate(0deg); }
+    25% { transform: rotate(-5deg); }
+    50% { transform: rotate(5deg); }
+    75% { transform: rotate(-5deg); }
+  }
+
+/* ==================================================
+================= MEDIA-QUERY-SECTION ===============
+=====================  768px  ==================== */
+
+@media (min-width: 48em) {
+
+    .typing-container {
+        font-size: 3rem;
+    }
+
+    .typing-text01 {
+        margin-left: 3.125rem;
+        margin-bottom: 0.625rem;
+    }
+
+    .notebook-image img {
+        max-width: 10rem;
+        margin-left: 4.375rem;
+    }
+
+    .book-image img {
+        width: 10rem;
+        height: 11.25rem;
+        margin-left: 3rem;
+        margin-top: 2rem;
+    }
+
+    .typing-effect-second {
+        margin-left: 6.25rem;
+    }
+
+    .typing-effect-third {
+        margin-left: 6rem;
+    }
+
+    .typing-effect-fourth {
+        margin-left: 3rem;
+    }
+
+    .typing-effect-fifth {
+        margin-left: 7rem;
+    }
+
+    .typing-effect::after,
+    .typing-effect-first::after,
+    .typing-effect-second::after,
+    .typing-effect-third::after,
+    .typing-effect-fourth::after,
+    .typing-effect-fifth::after {
+        height: 2.4rem;
+    }
+}
+
+/* ==================================================
+================= MEDIA-QUERY-SECTION ===============
+=====================  1024px  =================== */
+
+@media (min-width: 64em) {
+
+    .notebook-image img {
+        max-width: 14.375rem;
+        margin-left: 9.375rem;
+    }
+
+    .typing-container {
+        font-size: 4rem;
+    }
+
+    .notebook-image img {
+        max-width: 13.75rem;
+        margin-left: 5.625rem;
+    }
+
+    .book-image img {
+        width: 15.625rem;
+        height: 14.375rem;
+        margin-left: 3rem;
+        margin-top: 2rem;
+    }
+
+    .typing-effect::after,
+    .typing-effect-first::after,
+    .typing-effect-second::after,
+    .typing-effect-third::after,
+    .typing-effect-fourth::after,
+    .typing-effect-fifth::after {
+        height: 3.1rem;
+    }
+
+    .typing-effect-third {
+        margin-left: 6rem;
+    }
+
+    .typing-effect-fourth {
+        margin-left: 5rem;
+    }
+
+    .typing-effect-fifth {
+        margin-left: 10rem;
+    }
+}
+
+/* ==================================================
+================= MEDIA-QUERY-SECTION ===============
+=====================  1400px  =================== */
+
+@media (min-width: 87.5rem) {
+
+    .typing-container {
+        font-size: 5rem;
+    }
+
+    .typing-effect::after,
+    .typing-effect-first::after,
+    .typing-effect-second::after,
+    .typing-effect-third::after,
+    .typing-effect-fourth::after,
+    .typing-effect-fifth::after {
+        height: 3.8rem;
+    }
+}
+
+/* ==================================================
+================= MEDIA-QUERY-SECTION ===============
+=====================  1600px  =================== */
+
+@media (min-width: 100rem) {
+
+    .typing-container {
+        margin: 0 5rem;
+    }
+
+    .notebook-image img {
+        width: 18.75rem;
+        height: auto;
+    }
+
+    .typing-effect-first,
+    .typing-effect-second,
+    .typing-effect-third,
+    .typing-effect-fourth {
+        margin-bottom: 1.5rem;
+    }
+}
