+++ src/styles/components/worksection.css
@@ -1,0 +1,164 @@
+/* ========================================
+============== WORK-SECTION ==============
+======================================== */
+
+.works-section {
+    display: inline-block;
+    position: relative;
+    width: 100%;
+    margin: 0;
+    padding: 0;
+}
+
+.work {
+    display: block;
+    position: relative;
+    width: 100%;
+    height: 25rem;
+    background-position: top center;
+    background-size: cover;
+}
+
+.work-cover {
+    position: absolute;
+    top: 0;
+    left: 0;
+    width: 100%;
+    height: 100%;
+    background-color: black;
+    opacity: 0.5;
+}
+
+.work-text {
+    position: absolute;
+    top: 75%;
+    left: 0;
+    width: 100%;
+    text-align: center;
+}
+
+.work-text h2 {
+    text-transform: uppercase;
+    margin-top: 1rem;
+    margin-bottom: 0.5rem;
+}
+
+.work-text p {
+    font-family: var(--body-font);
+    text-align: center;
+    text-transform: uppercase;
+    margin-top: 0.625rem;
+    font-weight: bold;
+}
+
+#home-works .work-inner {
+    padding-top: 2.1875rem;
+    position: absolute;
+    top: 0;
+    left: 0;
+}
+
+.work-inner-wide {
+    font-size: 2rem;
+    text-transform: uppercase;
+    position: relative;
+    margin: 0 auto;
+    padding: 0 0.625rem;
+    max-width: 77.5rem;
+}
+
+.work01 {
+    background-image: url(/jv-background.png);
+    background-size: cover;
+    background-repeat: no-repeat;
+}
+
+.work02 {
+    background-image: url(/mv-background.jpg);
+    background-size: cover;
+    background-repeat: no-repeat;
+}
+
+.work03 {
+    background-image: url(/cap-background.jpg);
+    background-size: cover;
+    background-repeat: no-repeat;
+}
+
+.work04 {
+    background-image: url(/wallhaven.png);
+    background-size: cover;
+}
+
+/* ==================================================
+================= MEDIA-QUERY-SECTION ===============
+=====================  768px  ==================== */
+
+@media (min-width: 48em) {
+
+    .work-inner {
+        position: relative;
+        width: 100%;
+    }
+
+    .work-inner-wide h2 {
+        font-size: 2.5rem;
+        text-align: center;
+    }
+
+}
+
+/* ==================================================
+================= MEDIA-QUERY-SECTION ===============
+=====================  1024px  =================== */
+
+@media (min-width: 64em) {
+
+    .work-inner-wide h2 {
+        font-size: 3rem;
+        text-align: center;
+    }
+
+    #home-works {
+        height: 37.5rem;
+        margin: 0;
+        text-align: center;
+    }
+
+    .work {
+        position: absolute;
+        width: 25%;
+        height: 37.5rem;
+        background-position: center center;
+        background-size: cover;
+    }
+
+    .work01 {
+        top: 0;
+        left: 0;
+        background-color: hsl(25, 25, 25, 1);
+    }
+
+    .work02 {
+        top: 0;
+        left: 25%;
+        background-color: hsl(25, 25, 25, 1);
+    }
+
+    .work03 {
+        top: 0;
+        left: 50%;
+        background-color: hsl(25, 25, 25, 1);
+    }
+
+    .work04 {
+        top: 0;
+        left: 75%;
+        background-color: hsl(25, 25, 25, 1);
+    }
+
+    .work-inner {
+        margin: 1.5625rem auto;
+    }
+
+}
