+++ src/styles/ui/cursor.css
@@ -1,3 +1,7 @@
+/* ========================================
+============= CURSOR-STYLE ================
+======================================== */
+
 .cursor {
     pointer-events: none;
     position: fixed;
@@ -6,14 +10,14 @@
     mix-blend-mode: difference;
     top: 0;
     left: 0;
-    z-index: 9999999999999999;
+    z-index: 10000;
 }
 
 .circle {
     position: absolute;
     display: block;
-    width: 40px;
-    height: 40px;
-    border-radius: 20px;
+    width: 2.5rem;
+    height: 2.5rem;
+    border-radius: 1.25rem;
     background-color: #fff;
 }
